+++ windows/flutter/generated_plugin_registrant.cc
@@ -6,15 +6,12 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <bonsoir_windows/bonsoir_windows_plugin_c_api.h>
 #include <screen_retriever_windows/screen_retriever_windows_plugin_c_api.h>
 #include <tray_manager/tray_manager_plugin.h>
 #include <url_launcher_windows/url_launcher_windows.h>
 #include <window_manager/window_manager_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  BonsoirWindowsPluginCApiRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("BonsoirWindowsPluginCApi"));
   ScreenRetrieverWindowsPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("ScreenRetrieverWindowsPluginCApi"));
   TrayManagerPluginRegisterWithRegistrar(